Add gate-number to transcript on preprocessing
Implemented append to the transcript of the
gate-size.

Implemented tools to convert the gate size to
a byte slice in order to be able to generate a
`ff_fft::field::Fr` from bytes since otherways
we will need to use `BigInt` strures and import
more libs to the repo.

This aims to close #37