Implement PoseidonBranch & PoseidonLevel
The `PoseidonBranch` is the data structure over which
we will build our opening proofs.

- Implement `From<kelvin::Branch> for PoseidonBranch.
- Implement `PoseidonLevel` structure.
- Add docs for every structure.

These two allow us to obtain in the correct structure
the data obtained from a `kelvin` search to then elaborate
a merkle-opening proof.

We needed a way to collect the data from the kelvin `Branch` and
store it on a way which allows an easy Proof generation on later
stages.

The `From` trait implementation takes care of iterating over the Branch
levels and so, collect all of the data while also computes the bitflags
terms for each one of the leves. It also stores the index of the
position of the upper level hash to link the results and chain up the
hashing procedure until the root of the tree.

- Implemented also a few extra methods that allow to generate default
empty poseidon branches and levels so we have easier ways to generate
the `PoseidonBranch` in the `From` impl.

- Move ARITY definition to `lib.rs`.
- Add some module-docs.