Implement PoseidonAnnotation & Combine logic
In order to let kelvind understand how to build the `Brach`/path
we need indeed to implement `Combine` and `Content` traits for
a custom type which is `PoseidonAnnotation`.

This now guarantees that kelvin trees will know how to build the
path and provide it correctly to us.

- Implemented `Combine` for `PoseidonAnnotation`. This contains the
logic for hashing an entire merkle tree level computing the bitflags
and then, returning the resulting value located on the second leaf.

- Implemented `Content` for `PoseidonAnnotation` since it's required
directly by kelvin.

Since it's needed to know when a leaf is empty, we need to
use `Option<T>` instead of `T` directly.

- Added lib exported structs and docs.

Closes #12