Fix SPAG for docs (#254)
* Render Commitment docs

Previosuly the commitment scheme docs would not render in the KaTeX format. The docs have been checked against a local `index.html`, to show that they render correctly.

* Rendering SNARK notes

Previosuly the SNARK docs would not render in the KaTeX format. The docs have been checked against a local `index.html`, to show that they render correctly.

* Fix typo

Fixing typo/format of target polynomial

* Rendering proof notes

Previosuly the PROOF docs would not render in the KaTeX format. The docs have been checked against a local `index.html`, to show that they render correctly.

* Fix SPAG

The Spelling Punctuation And Grammar required a review on the docs PR as part of a tidy-up. This commit should remove all SPAG errors