Update `dusk-plonk` version to v0.2.7
This new version comes with a lot of new features. But none of them
really impacts this library. But it does impact the libraries that use
this one as a dependency.

Therefore, we update the version of plonk to avoid dependency issues.

- Replaced `add_constant_witness` function by
`add_witness_to_circuit_description` since the first one has been
deprecated.

Closes #46