Refactor padding for PoseidonBranch
We realized that there was a problem with the previous extend
approach used by the PoseidonBranch.
It was extending the branch on a way on which it was also forced to
modigy the root. The problem is obvious, and it's that the Verifier
has the original root, not the extended one (which doesn't even exist).

Therefore, we've refactored the extension and it's just only applied to
the gadget opening, not to the scalar one since it's not important at all
to pad there (we don't care about constant-sice branches on the "out of
circuits" domain.

- Removed extension fn's for the gadget and the scalar approaches.
- Refactored the padding in order to be performed on a different way:
We now load into the PoseidonBranch the levels required to pad the
branch and they're set to zero. Once this is done, inside of the gadget,
we hash them without any other purpose rather than obtain the same
ammount of constrats for all of the circuits. We can understand this
padding as an addition of dummy constraints.
- Refactored From<&Branch> for PoseidonBranch impl.
- Added a `root()` method for branch and made it's internals
`pub(crate)`.
- Refactored `merkle_opening_gadget` to perform the dummy constraints as
padding inside.

Closes #71