Fix PR from review
As per the most recent review
a few different things have
been changed on this PR.

The first is the use of extended points
as the representation. Using a function
to create the hash challenge is also
added. Using a random secret for the
failure tests. Renaming the tests file.