Add gate-featured `Canon` impl/usage
Since we don't necessarily need `Canon` derived for
our structures. Therefore, a gate-featured solution for
canonnical and it's deps is the best solution.

Therefore, `Canon` is not derived for `PoseidonCipher` unless
we use the `canon` feature.
Same happens for the `PoseidonTree`. The tree module will not be
exported if the `canon` feature is not used.