Re-design the ecc module API
The module contained a lot of functions that could be associated
to the `StandardComposer` directly but for some reason, they weren't,
which forced the users to need to import manually the functions from
the paths.
Also some of the gadgets implemented for the `Point` & `PointScalar`
structures weren't also avaliable or not associated with the
`StarndardComposer` which is the default behaviour with the normal
`Variable` API.

- Moved all gadget-related associated functions to be associated
methods of `StandardComposer`.
- Moved the `Point` generation fn's to also be `StandardComposer`
associated methods.
- Renamed some of the methods to make them more clear and inline
with the `Variable` API.
- Removed duplicities in the functions implemented. For example
addition was implemented for `Point` and `StandardComposer` at the
same time. It's better to provide a single way of doing things so
circuits are easier to read.

Resolves: #456