Remove Makefile & docs.rs metadata
Since we don't publish our docs in docs.rs with the `nightly` feature
enabled and this one is the only one that requires us to make usage
of KateX, we've simply removed the Makefile and also the metadata from
docs.rs since it was useless.

From now on, users will be able to generate extensive docs with KateX
rendering still, but running the appropiated command listed in README.md.