Add getters to avoid trimming for OpeningKey obtention
There are times as for example in the `Circuit::verify_proof` fn
where we just need to retrive the `OpeningKey` from a
`PublicParameters` instance without any need of trimming them
ie. trim the `CommitKey`.

Therefore, this PR adds an alternative way to obtain the `CommitKey`
and the `OpeningKey` and implements the second one in the circuit
module.

Resolves: #510